+++ src/utils/actionHandler.ts
@@ -1563,7 +1563,7 @@ export async function getTicker(name, chain_name = null) {
           { subscribe: false, params: { chain_name, hash: name.split('/')[1] } },
           { root: true },
         ));
-      return await getDisplayName(verifyTrace.base_denom);
+      return await getTicker(verifyTrace.base_denom);
     } catch (e) {
       console.error(e);
       return name + '(unverified)';