+++ src/components/common/TotalLiquidityPrice.vue
@@ -1,17 +1,18 @@
 <template>
-  {{ totalLiquidityPrice ? toUSD(totalLiquidityPrice) : '-' }}
+  <CurrencyDisplay :value="totalLiquidityPrice" show-dash />
 </template>
 
 <script lang="ts">
 import { defineComponent, PropType } from 'vue';
 
+import CurrencyDisplay from '@/components/ui/CurrencyDisplay.vue';
 import usePool from '@/composables/usePool';
 import { Pool } from '@/types/actions';
 
 //import TrendingUpIcon from '../common/Icons/TrendingUpIcon.vue';
 
 export default defineComponent({
-  name: 'TotalLiquidityPrice',
+  components: { CurrencyDisplay },
 
   //components: { TrendingUpIcon },
 
@@ -23,21 +24,9 @@ export default defineComponent({
   },
 
   setup(props) {
-    const toUSD = (value) => {
-      var formatter = new Intl.NumberFormat('en-US', {
-        style: 'currency',
-        currency: 'USD',
-
-        // These options are needed to round to whole numbers if that's what you want.
-        //minimumFractionDigits: 0, // (this suffices for whole numbers, but will print 2500.10 as $2,500.1)
-        //maximumFractionDigits: 0, // (causes 2500.99 to be printed as $2,501)
-      });
-      return formatter.format(value);
-    };
-
     const { totalLiquidityPrice } = usePool(props.pool.id);
 
-    return { totalLiquidityPrice, toUSD };
+    return { totalLiquidityPrice };
   },
 });
 </script>