+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Fixes
 
 - Improved partial swap result
+- Fixed pool price display decimals
 
 ## 0.1.2 - 2021-09-29