+++ CHANGELOG.md
@@ -8,10 +8,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Updates
 
+- Improve asset order in Receive page
+
 ### Fixes
 
 - Improved partial swap result
 - Fixed ability to change chain on Withdraw page
+
+### Perfomance
 
 ## 0.1.2 - 2021-09-29