+++ CHANGELOG.md
@@ -16,11 +16,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Improved partial swap result
 - Fixed pool price display decimals
 - Fixed ability to change chain on Withdraw page
+- Fixed No pool warning in Swap
 
 ### Perfomance
 
 ### Perfomance
-
 
 ## 0.1.2 - 2021-09-29